Splits parity node into two and fixes both.
A handful of minor issues with the Parity node, and also splits it in two.  Along the way I removed the option from the Parity node to choose sealing time, we never use it and it adds complexity.  Adding it back should we want it is easy.  I didn't address the issue with sharing the `common_start.sh` across folders, as I'm not sure what the best way to do that is.